build: ensure consistent versioning
The version is set in config.mk.in which is transformed to config.mk by
configure.
The Problem: If we already have a configured local copy and we pull a version
change (in config.mk.in), a following build (without calling configure) will
lead to a wrong built-in version.
We fix this by
 1. moving the version inside the Makefile
 2. ensure that the all objs are rebuilt if the Makefile was changed